pyproject.toml
[project]
name = "xcnt-expense-manager"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "django",
    "django-ninja",
    "uvicorn[standard]",
    "requests",
    "psycopg2-binary",
    "mypy",
    "django-stubs[compatible-mypy]==5.1.3",
    "pytest",
    "pytest-django",
    "factory-boy",
    "Faker",
    "types-requests",
    "ruff==0.7.4",
    "import-linter==2.1",
    # Typing stubs
    "types-python-dateutil",
    "types-jmespath",
    "types-oauthlib",
    "types-requests-oauthlib",
    "pydantic-settings"
]

[tool.setuptools]
# Let setuptools find packages automatically

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "xcnt.settings"
python_files = ["test_*.py", "*_test.py"]

[tool.mypy]
files = "."
exclude = ["build"]
ignore_missing_imports = true
mypy_path = "$MYPY_CONFIG_FILE_DIR/.:$MYPY_CONFIG_FILE_DIR/stubs"
plugins = [
    "mypy_django_plugin.main",
    "pydantic.mypy",
]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_subclassing_any = false
disable_error_code = ["no-untyped-def"]

[[tool.mypy.overrides]]
module = "tests.factories.*"
disallow_subclassing_any = false

[[tool.mypy.overrides]]
module = [
    "uuid_extensions",
    "factory",
    "pydantic-settings"
]

[tool.django-stubs]
django_settings_module = "xcnt.settings"

[tool.pydantic-mypy]
init_typed = true
init_forbid_extra = true